+++ koku/api/report/aws/queries.py
@@ -18,7 +18,7 @@
 import copy
 import datetime
 import logging
-from collections import OrderedDict, UserDict, defaultdict
+from collections import OrderedDict
 from decimal import Decimal, DivisionByZero, InvalidOperation
 from itertools import groupby
 
@@ -35,6 +35,7 @@ from django.db.models.functions import (Concat,
                                         TruncMonth)
 from tenant_schemas.utils import tenant_context
 
+from api.report.query_filter import QueryFilter, QueryFilterCollection
 from api.utils import DateHelper
 from reporting.models import (AWSCostEntryLineItem,
                               AWSCostEntryLineItemAggregates,
@@ -64,7 +65,7 @@ class ProviderMap(object):
     data - counts, costs, etc. The only variable is determining which
     DB tables and fields are supplying the requested data.
 
-    ProviderMap supplies ReportQueryHandler with the appropriate model
+    ProviderMap supplies AWSReportQueryHandler with the appropriate model
     references.
     """
 
@@ -237,257 +238,7 @@ class TruncMonthString(TruncMonth):
         return value.strftime('%Y-%m')
 
 
-class QueryFilter(UserDict):
-    """Dict-like object representing a single query filter."""
-
-    SEP = '__'    # separator
-    table = None
-    field = None
-    operation = None
-    parameter = None
-
-    def __init__(self, table=None, field=None, operation=None, parameter=None):
-        """Constructor.
-
-        Args:
-            table (str) - The name of a DB table
-            field (str) - The name of a DB field
-            operation (str) - The name of a DB operation, e.g. 'in' or 'gte'
-            parameter (object) - A valid query target, e.g. a list or datetime
-        """
-        super().__init__(table=table, field=field, operation=operation,
-                         parameter=parameter)
-        self.table = table
-        self.field = field
-        self.operation = operation
-        self.parameter = parameter
-
-    def composed_query_string(self):
-        """Return compiled query string."""
-        fields = [entry for entry in [self.table, self.field, self.operation]
-                  if entry is not None]
-        return self.SEP.join(fields)
-
-    def composed_Q(self):
-        """Return a Q object formatted for Django's ORM."""
-        query_dict = {self.composed_query_string(): self.parameter}
-        return Q(**query_dict)
-
-    def from_string(self, query_string):
-        """Parse a string representing a filter.
-
-        Returns:
-            QueryFilter instance.
-
-        Args:
-            query_string (str) A string representing a query filter.
-
-        Example:
-            QueryFilter().from_string('mytable__myfield__contains')
-
-        """
-        parts = query_string.split(self.SEP)
-        if len(parts) == 3:
-            self.table, self.field, self.operation = parts
-        elif len(parts) == 2:
-            self.table, self.operation = parts
-        else:
-            message = 'Incorrect number of parts in query string. ' + \
-                'Need at least two of [table, field, operation].'
-            raise TypeError(message)
-        return self
-
-    def __eq__(self, other):
-        """Exact comparison."""
-        return self.data == other.data
-
-    def __repr__(self):
-        """Return string representation."""
-        return str(self.composed_Q())
-
-
-class QueryFilterCollection(object):
-    """Object representing a set of filters for a query.
-
-    This object behaves in list-like ways.
-    """
-
-    def __init__(self, filters=None):
-        """Constructor.
-
-        Args:
-            filters (list) a list of QueryFilter instances.
-        """
-        if filters is None:
-            self._filters = list()    # a list of QueryFilter objects
-        else:
-            if not isinstance(filters, list):
-                raise TypeError('filters must be a list')
-
-            if not all(isinstance(item, QueryFilter) for item in filters):
-                raise TypeError('Filters list must contain QueryFilters.')
-
-            self._filters = filters
-
-    def add(self, query_filter=None, table=None, field=None,
-            operation=None, parameter=None):
-        """Add a query filter to the collection.
-
-        QueryFilterCollection does try to maintain filter uniqueness. A new
-        object will not be added to the collection if it already exists.
-
-        Args:
-            query_filter (QueryFilter) a QueryFilter object
-
-            - or -
-
-            table (str)  db table name
-            field (str)  db field/row name
-            operation (str) db operation
-            parameter (object) query object
-
-        """
-        error_message = 'query_filter can not be defined with other parameters'
-        if query_filter and (table or field or operation or parameter):
-            raise AttributeError(error_message)
-
-        if query_filter and query_filter not in self:
-            self._filters.append(query_filter)
-
-        if (table or field or operation or parameter):
-            qf = QueryFilter(table=table, field=field, operation=operation,
-                             parameter=parameter)
-            if qf not in self:
-                self._filters.append(qf)
-
-    def compose(self):
-        """Compose filters into a dict for submitting to Django's ORM."""
-        composed_query = None
-        compose_dict = defaultdict(list)
-        for filt in self._filters:
-            filt_key = filt.composed_query_string()
-            compose_dict[filt_key].append(filt)
-
-        for filter_list in compose_dict.values():
-            or_filter = None
-            for filter_item in filter_list:
-                if or_filter is None:
-                    or_filter = filter_item.composed_Q()
-                else:
-                    or_filter = or_filter | filter_item.composed_Q()
-            if composed_query is None:
-                composed_query = or_filter
-            else:
-                composed_query = composed_query & or_filter
-        return composed_query
-
-    def __contains__(self, item):
-        """Return a boolean about whether `item` is in the collection.
-
-        This will do both an exact and a fuzzy match. Exact matches are
-        preferred over fuzzy matches. However, if there is no exact match,
-        __contains__ will return `True` if there is an object that contains all
-        of the same properties as `item`, even if additional properties are
-        set on the object in our collection. (See: `QueryFilterCollection.get()`)
-
-        Args:
-            item (QueryFilter or dict) object to search for.
-
-        Returns:
-            (bool) Whether a matching object was found.
-
-        """
-        if isinstance(item, QueryFilter):
-            return item in self._filters
-        if isinstance(item, dict) and self.get(item):
-            return True
-        return False
-
-    def __iter__(self):
-        """Return an iterable of the collection."""
-        return self._filters.__iter__()
-
-    def __getitem__(self, key):
-        """Return object identified by key.
-
-        Args:
-            key (int) the index of the QueryFilter to return.
-        """
-        return self._filters[key]
-
-    def __repr__(self):
-        """Return string representation."""
-        out = f'{self.__class__}: '
-        for filt in self._filters:
-            out += filt.__repr__() + ', '
-        return out
-
-    def delete(self, query_filter=None, table=None, field=None,
-               operation=None, parameter=None):
-        """Delete a query filter from the collection.
-
-        Args:
-            query_filter (QueryFilter) a QueryFilter object
-
-            - or -
-
-            table (str)  db table name
-            field (str)  db field/row name
-            operation (str) db operation
-            parameter (object) query object
-
-        """
-        error_message = 'query_filter can not be defined with other parameters'
-        if query_filter and (table or field or operation or parameter):
-            raise AttributeError(error_message)
-
-        if query_filter and query_filter in self:
-            self._filters.remove(query_filter)
-
-        if (table or field or operation or parameter):
-            qf = QueryFilter(table=table, field=field, operation=operation,
-                             parameter=parameter)
-            if qf in self:
-                self._filters.remove(qf)
-
-    def get(self, search):
-        """Retrieve the first matching filter in the collection.
-
-        This is a "fuzzy" search. This method looks for an object that matches
-        all key-value pairs of `search` with an object in the collection. The
-        matched object may have additional properties set to a non-None value.
-
-        Args:
-            search (dict) A dictionary of filter parameters to search for.
-
-        Example:
-
-            These examples are in order of least-specific (most likely to have
-            more than one result) to most-specific (most likely to have exactly
-            one result).
-
-            QueryFilterCollection.get({'operation': 'contains'})
-
-            QueryFilterCollection.get({'table': 'mytable', 'operation': 'gte'})
-
-            QueryFilterCollection.get({'table': 'mytable',
-                                       'field': 'myfield',
-                                       'operation': 'in'})
-
-        """
-        for idx, filt in enumerate(self._filters):
-            filter_values = [filt.get(key) for key in search.keys()
-                             if filt.get(key)]
-            search_values = [search.get(key) for key in search.keys()
-                             if search.get(key)]
-            filter_values.sort()
-            search_values.sort()
-            if search_values == filter_values:
-                return (idx, filt)
-        return None
-
-
-class ReportQueryHandler(object):
+class AWSReportQueryHandler(object):
     """Handles report queries and responses."""
 
     default_ordering = {'total': 'desc'}
@@ -762,7 +513,7 @@ class ReportQueryHandler(object):
             group_by = self.get_query_param_data('group_by', q_param, list())
             filter_ = self.get_query_param_data('filter', q_param, list())
             list_ = list(set(group_by + filter_))    # uniquify the list
-            if list_ and not ReportQueryHandler.has_wildcard(list_):
+            if list_ and not AWSReportQueryHandler.has_wildcard(list_):
                 for item in list_:
                     q_filter = QueryFilter(parameter=item, **filt)
                     filters.add(q_filter)
@@ -872,9 +623,9 @@ class ReportQueryHandler(object):
 
         for key, group in groupby(data, lambda by: by.get(curr_group)):
             grouped = list(group)
-            grouped = ReportQueryHandler._group_data_by_list(group_by_list,
-                                                             (group_index + 1),
-                                                             grouped)
+            grouped = AWSReportQueryHandler._group_data_by_list(group_by_list,
+                                                                (group_index + 1),
+                                                                grouped)
             datapoint = out_data.get(key)
             if datapoint and isinstance(datapoint, dict):
                 out_data[key].update(grouped)
@@ -946,8 +697,8 @@ class ReportQueryHandler(object):
             if self._limit:
                 data = self._ranked_list(data_list)
             group_by = self._get_group_by()
-            grouped = ReportQueryHandler._group_data_by_list(group_by, 0,
-                                                             data)
+            grouped = AWSReportQueryHandler._group_data_by_list(group_by, 0,
+                                                                data)
             bucket_by_date[date] = grouped
         return bucket_by_date
